casino-bot/requirements.txt
pyautogui>=0.9.54
opencv-python>=4.9.0
Pillow>=10.2.0
pytesseract>=0.3.10
PyYAML>=6.0.1
numpy>=1.26.0
